+++ src/js/components/sections/DetailsSection.jsx
