+++ src/js/components/sections/MajorDivider.jsx
